Spiritual Development v0.1
This is the first iteration of it using the NER dictionary format to search and the new WB (WordBank) format to store the dictionary. This version takes the words you enter in the search bar and splits them into words then looks for each word in each verse. This version scans all 31K verses but it is actually quite quick. The next iteration will start to use synonyms and then in the third iteration will use the recognition provider.  This version is useful to find a verse if you can think of a few words in it.